Fixed floating point bug with calculator addition.
Calculator now properly handles additions like 2.5 + 3.

+++ calc/math_lib.js
@@ -4,7 +4,7 @@ module.exports = {
 };
 
 function add(a, b) {
-    return parseInt(a) + parseInt(b);
+    return parseFloat(a) + parseFloat(b);
 };
 
 function mul(a, b) {
